move the (backend) application to a (newly-created) package
