+++ .vscode/launch.json
@@ -11,7 +11,7 @@
       "cwd": "${workspaceFolder}/backend",
       "module": "flask",
       "env": {
-        "FLASK_APP": "vocab_treasury.py",
+        "FLASK_APP": "src",
         "FLASK_ENV": "development"
       },
       "args": ["run", "--no-debugger"],
@@ -22,7 +22,7 @@
       "type": "python",
       "request": "launch",
       "cwd": "${workspaceFolder}/backend",
-      "module": "vocab_treasury",
+      "module": "run_dev_server",
       "env": {
         "FLASK_ENV": "development"
       },
