+++ backend/run_dev_server.py
@@ -1,0 +1,5 @@
+from src.vocab_treasury import app
+
+
+if __name__ == "__main__":
+    app.run(use_debugger=False, use_reloader=False, passthrough_errors=True)
