+++ backend/src/__init__.py
@@ -1,0 +1,4 @@
+from src.vocab_treasury import app, db
+
+
+__all__ = [app, db]
